# go-microservices/products/swagger.yaml
basePath: /
consumes:
    - application
definitions:
    Product:
        description: Product Product product
        properties:
            description:
                description: description
                type: string
                x-go-name: Description
            id:
                description: the id for this product
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            name:
                description: name
                type: string
                x-go-name: Name
            price:
                description: price
                format: float
                type: number
                x-go-name: Price
            sku:
                description: s k u
                type: string
                x-go-name: SKU
        required:
            - id
        type: object
        x-go-package: microservices/models
info:
    description: '# Documentation for Product API'
    title: of Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Returns a list of products
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
    /products/{id}:
        delete:
            description: deletes a product
            operationId: deleteProduct
            parameters:
                - description: The id of the product to delete from database
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContent'
            tags:
                - products
produces:
    - application/json
responses:
    noContent:
        description: ""
    productsResponse:
        description: ""
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
schemes:
    - http
swagger: "2.0"